pyproject.toml
[tool.poetry]
name = "valorant-stats"
version = "0.1.0"
description = "Personal Valorant statistics dashboard"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
django-ninja-extra = "^0.19.1"
pymongo = "^4.6.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
auth0-python = "^4.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
